Refactor: Change order_regels person_id to unsignedInteger

+++ database/migrations/2025_10_14_120000_add_person_id_to_orderregels_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('order_regels', function (Blueprint $table) {
-            $table->unsignedBigInteger('person_id')->nullable()->after('product_id');
+            $table->unsignedInteger('person_id')->nullable()->after('product_id');
             $table->foreign('person_id')->references('id')->on('persons')->onDelete('set null');
         });
     }
